+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add sizes of data import and fetcher queues to instance report.
+
 ### Changed
 
 - Increased incoming activity queue batch size to 20.